README.md
# SkillsClub
This is an android app for my college SkillsClub. Using this android app students can able to register in the events conducting by the club. There are separate views for both student and admin.Admin login has provided with features like adding an event, disenrolling the students,deleting the events,sending notifications to students.